refactor: split prefer-source-imports helpers
