+++ src/rules/prefer-source-imports.ts
@@ -1,494 +1,50 @@
-/// <reference types="node" />
 import fs from 'node:fs';
 import path from 'node:path';
 import { AST_NODE_TYPES, TSESLint, TSESTree } from '@typescript-eslint/utils';
-// eslint-disable-next-line @typescript-eslint/no-require-imports
-const parser = require('@typescript-eslint/parser');
+import {
+  collectAllExportedBindings,
+  collectBarrelAnalysis,
+  collectExportedBindings,
+  getReExportKey,
+  getReExportMeta,
+  parseBarrelFile,
+  parseExportedBindings,
+  parseModule,
+  resolveReExportTarget,
+} from './prefer-source-imports/analysis';
+import {
+  buildAutofix,
+  getMergeableImportDeclaration,
+  isTypeOnlyImport,
+  serializeImportBinding,
+  serializeImportBindings,
+} from './prefer-source-imports/autofix';
+import {
+  applyAliasTarget,
+  buildAliasSpecifier,
+  getOptionsCacheKey,
+  isRelativePath,
+  matchesAliasPattern,
+  normalizeModulePath,
+  resolveModuleFile,
+  toRelativeImportSpecifier,
+} from './prefer-source-imports/path-utils';
+import {
+  BarrelAnalysis,
+  ManualAliasMapping,
+  MatchedSpecifier,
+  MessageIds,
+  NamedImportSpecifier,
+  Options,
+  ReExportTarget,
+  TsconfigInfo,
+} from './prefer-source-imports/types';
 // eslint-disable-next-line @typescript-eslint/no-require-imports
 const typescript = require('typescript');
 
-type MessageIds = 'preferSourceImport' | 'preferSourceImports';
-type Options = [
-  {
-    fixStyle?: 'auto' | 'preserve-alias' | 'relative';
-    paths?: Record<string, string | string[]>;
-    tsconfig?: boolean | string;
-  }?,
-];
-
-type ReExportTarget = {
-  importedName: string;
-  resolvedFilePath: string;
-  sourceSpecifier: string;
-  isTypeOnly: boolean;
-  fromExportAll: boolean;
-};
-
-type NamedImportSpecifier = TSESTree.ImportSpecifier & {
-  imported: TSESTree.Identifier;
-  local: TSESTree.Identifier;
-};
-
-type MatchedSpecifier = {
-  preferredSourceSpecifier: string | null;
-  specifier: NamedImportSpecifier;
-  reExportTarget: ReExportTarget;
-};
-
-type BarrelAnalysis = {
-  explicitReExports: Map<string, ReExportTarget>;
-  exportAllReExports: Map<string, ReExportTarget>;
-};
-
-const SOURCE_FILE_EXTENSIONS = ['.ts', '.tsx', '.js', '.jsx', '.mts', '.cts', '.mjs', '.cjs'] as const;
-
-type TsconfigInfo = {
-  compilerOptions: Record<string, unknown>;
-  configFilePath: string;
-};
-
-type ManualAliasMapping = {
-  pattern: string;
-  target: string;
-};
-
-const parsedModuleCache = new Map<string, TSESTree.Program | null>();
-const exportedBindingsCache = new Map<string, Set<string>>();
-const moduleFileResolutionCache = new Map<string, string | null>();
 const importResolutionCache = new Map<string, string | null>();
 const barrelAnalysisCache = new Map<string, BarrelAnalysis | null>();
 const tsconfigInfoCache = new Map<string, TsconfigInfo | null>();
-
-function getOptionsCacheKey(options: Options[0] | undefined, cwd: string): string {
-  return JSON.stringify({
-    cwd,
-    options: options ?? {},
-  });
-}
-
-function isRelativePath(value: string): boolean {
-  return value.startsWith('.');
-}
-
-function getReExportKey(exportedName: string, isTypeOnly: boolean): string {
-  return `${isTypeOnly ? 'type' : 'value'}:${exportedName}`;
-}
-
-function getReExportMeta(reExportKey: string): { exportedName: string; isTypeOnly: boolean } {
-  if (reExportKey.startsWith('type:')) {
-    return {
-      exportedName: reExportKey.slice(5),
-      isTypeOnly: true,
-    };
-  }
-
-  return {
-    exportedName: reExportKey.slice(6),
-    isTypeOnly: false,
-  };
-}
-
-function resolveModuleFile(importerFilename: string, specifier: string): string | null {
-  const cacheKey = `${importerFilename}\0${specifier}`;
-  const cachedResult = moduleFileResolutionCache.get(cacheKey);
-
-  if (cachedResult !== undefined) {
-    return cachedResult;
-  }
-
-  const resolvedBase = path.resolve(path.dirname(importerFilename), specifier);
-  const candidatePaths = [
-    resolvedBase,
-    ...SOURCE_FILE_EXTENSIONS.map(extension => `${resolvedBase}${extension}`),
-    ...SOURCE_FILE_EXTENSIONS.map(extension => path.join(resolvedBase, `index${extension}`)),
-  ];
-
-  const resolvedFilePath =
-    candidatePaths.find(candidatePath => fs.existsSync(candidatePath) && fs.statSync(candidatePath).isFile()) ?? null;
-
-  moduleFileResolutionCache.set(cacheKey, resolvedFilePath);
-
-  return resolvedFilePath;
-}
-
-function normalizeModulePath(filePath: string): string {
-  const normalizedPath = filePath.split(path.sep).join('/');
-
-  return normalizedPath.replace(/\/index(?=(\.[^./]+)?$)/, '').replace(/\.[^./]+$/, '');
-}
-
-function toRelativeImportSpecifier(importerFilename: string, resolvedFilePath: string): string {
-  const relativePath = path.relative(path.dirname(importerFilename), resolvedFilePath);
-  const normalizedPath = normalizeModulePath(relativePath);
-
-  return normalizedPath.startsWith('.') ? normalizedPath : `./${normalizedPath}`;
-}
-
-function matchesAliasPattern(specifier: string, pattern: string): string | null {
-  if (!pattern.includes('*')) {
-    return specifier === pattern ? '' : null;
-  }
-
-  const [rawPrefix = '', rawSuffix = ''] = pattern.split('*');
-  const prefix = rawPrefix;
-  const suffix = rawSuffix;
-
-  if (!specifier.startsWith(prefix) || !specifier.endsWith(suffix)) {
-    return null;
-  }
-
-  return specifier.slice(prefix.length, specifier.length - suffix.length);
-}
-
-function applyAliasTarget(target: string, wildcardValue: string): string {
-  return target.includes('*') ? target.replace('*', wildcardValue) : target;
-}
-
-function buildAliasSpecifier(pattern: string, wildcardValue: string): string {
-  return pattern.includes('*') ? pattern.replace('*', wildcardValue) : pattern;
-}
-
-function parseModule(filePath: string): TSESTree.Program | null {
-  const cachedProgram = parsedModuleCache.get(filePath);
-
-  if (cachedProgram !== undefined) {
-    return cachedProgram;
-  }
-
-  let sourceText: string;
-
-  try {
-    sourceText = fs.readFileSync(filePath, 'utf8');
-  } catch {
-    parsedModuleCache.set(filePath, null);
-    return null;
-  }
-
-  try {
-    const program = parser.parse(sourceText, {
-      filePath,
-      ecmaVersion: 'latest',
-      sourceType: 'module',
-      range: false,
-      loc: false,
-      comment: false,
-      tokens: false,
-    }) as TSESTree.Program;
-
-    parsedModuleCache.set(filePath, program);
-
-    return program;
-  } catch {
-    parsedModuleCache.set(filePath, null);
-    return null;
-  }
-}
-
-function collectExportedBindings(program: TSESTree.Program): Set<string> {
-  const exportedBindings = new Set<string>();
-
-  function addExportedBinding(exportedName: string, isTypeOnly: boolean): void {
-    exportedBindings.add(getReExportKey(exportedName, isTypeOnly));
-  }
-
-  program.body.forEach(statement => {
-    if (statement.type === AST_NODE_TYPES.ExportNamedDeclaration) {
-      if (statement.declaration) {
-        switch (statement.declaration.type) {
-          case AST_NODE_TYPES.VariableDeclaration:
-            statement.declaration.declarations.forEach(declaration => {
-              if (declaration.id.type === AST_NODE_TYPES.Identifier) {
-                addExportedBinding(declaration.id.name, false);
-              }
-            });
-            break;
-          case AST_NODE_TYPES.FunctionDeclaration:
-            if (statement.declaration.id) {
-              addExportedBinding(statement.declaration.id.name, false);
-            }
-            break;
-          case AST_NODE_TYPES.ClassDeclaration:
-          case AST_NODE_TYPES.TSEnumDeclaration:
-            if (statement.declaration.id) {
-              addExportedBinding(statement.declaration.id.name, false);
-              addExportedBinding(statement.declaration.id.name, true);
-            }
-            break;
-          case AST_NODE_TYPES.TSInterfaceDeclaration:
-          case AST_NODE_TYPES.TSTypeAliasDeclaration:
-            if (statement.declaration.id) {
-              addExportedBinding(statement.declaration.id.name, true);
-            }
-            break;
-        }
-      }
-
-      statement.specifiers.forEach(specifier => {
-        if (specifier.type !== AST_NODE_TYPES.ExportSpecifier) {
-          return;
-        }
-
-        if (specifier.exported.type === AST_NODE_TYPES.Identifier) {
-          addExportedBinding(
-            specifier.exported.name,
-            statement.exportKind === 'type' || specifier.exportKind === 'type',
-          );
-        }
-      });
-    }
-  });
-
-  return exportedBindings;
-}
-
-function parseExportedBindings(filePath: string): Set<string> {
-  const cachedBindings = exportedBindingsCache.get(filePath);
-
-  if (cachedBindings) {
-    return cachedBindings;
-  }
-
-  const program = parseModule(filePath);
-
-  if (!program) {
-    const emptyBindings = new Set<string>();
-    exportedBindingsCache.set(filePath, emptyBindings);
-
-    return emptyBindings;
-  }
-
-  const exportedBindings = collectExportedBindings(program);
-  exportedBindingsCache.set(filePath, exportedBindings);
-
-  return exportedBindings;
-}
-
-function collectAllExportedBindings(
-  filePath: string,
-  resolveImport: (importerFilename: string, specifier: string) => string | null,
-  visitedFiles: Set<string> = new Set(),
-): Set<string> {
-  if (visitedFiles.has(filePath)) {
-    return new Set();
-  }
-
-  const program = parseModule(filePath);
-
-  if (!program) {
-    return new Set();
-  }
-
-  const nextVisitedFiles = new Set(visitedFiles);
-  nextVisitedFiles.add(filePath);
-
-  const exportedBindings = collectExportedBindings(program);
-
-  program.body.forEach(statement => {
-    if (
-      statement.type !== AST_NODE_TYPES.ExportAllDeclaration ||
-      statement.exported ||
-      statement.source.type !== AST_NODE_TYPES.Literal ||
-      typeof statement.source.value !== 'string'
-    ) {
-      return;
-    }
-
-    const resolvedSourceFile = resolveImport(filePath, statement.source.value);
-
-    if (!resolvedSourceFile) {
-      return;
-    }
-
-    const nestedBindings = collectAllExportedBindings(resolvedSourceFile, resolveImport, nextVisitedFiles);
-
-    nestedBindings.forEach(reExportKey => {
-      const { isTypeOnly } = getReExportMeta(reExportKey);
-
-      if (statement.exportKind === 'type' && !isTypeOnly) {
-        return;
-      }
-
-      if (exportedBindings.has(reExportKey)) {
-        return;
-      }
-
-      exportedBindings.add(reExportKey);
-    });
-  });
-
-  return exportedBindings;
-}
-
-function resolveReExportTarget(
-  filePath: string,
-  exportedName: string,
-  isTypeOnly: boolean,
-  resolveImport: (importerFilename: string, specifier: string) => string | null,
-  visitedTargets: Set<string> = new Set(),
-): ReExportTarget | null {
-  const reExportKey = getReExportKey(exportedName, isTypeOnly);
-  const visitedTargetKey = `${filePath}:${reExportKey}`;
-
-  if (visitedTargets.has(visitedTargetKey)) {
-    return null;
-  }
-
-  const program = parseModule(filePath);
-
-  if (!program) {
-    return null;
-  }
-
-  const nextVisitedTargets = new Set(visitedTargets);
-  nextVisitedTargets.add(visitedTargetKey);
-
-  const barrelAnalysis = collectBarrelAnalysis(program, filePath, resolveImport, nextVisitedTargets);
-
-  if (!barrelAnalysis) {
-    return null;
-  }
-
-  return (
-    barrelAnalysis.explicitReExports.get(reExportKey) ?? barrelAnalysis.exportAllReExports.get(reExportKey) ?? null
-  );
-}
-
-function collectBarrelAnalysis(
-  program: TSESTree.Program,
-  barrelFilePath: string,
-  resolveImport: (importerFilename: string, specifier: string) => string | null,
-  visitedTargets: Set<string> = new Set(),
-): BarrelAnalysis | null {
-  const explicitReExports = new Map<string, ReExportTarget>();
-  const exportAllReExports = new Map<string, ReExportTarget>();
-
-  program.body.forEach(statement => {
-    const source = statement.type === AST_NODE_TYPES.ExportNamedDeclaration ? statement.source : null;
-
-    if (
-      statement.type === AST_NODE_TYPES.ExportNamedDeclaration &&
-      source?.type === AST_NODE_TYPES.Literal &&
-      typeof source.value === 'string'
-    ) {
-      const resolvedSourceFile = resolveImport(barrelFilePath, source.value);
-
-      if (!resolvedSourceFile) {
-        return;
-      }
-
-      statement.specifiers.forEach(specifier => {
-        if (
-          specifier.type !== AST_NODE_TYPES.ExportSpecifier ||
-          specifier.local.type !== AST_NODE_TYPES.Identifier ||
-          specifier.exported.type !== AST_NODE_TYPES.Identifier
-        ) {
-          return;
-        }
-
-        const specifierIsTypeOnly = statement.exportKind === 'type' || specifier.exportKind === 'type';
-        const resolvedTarget =
-          resolveReExportTarget(
-            resolvedSourceFile,
-            specifier.local.name,
-            specifierIsTypeOnly,
-            resolveImport,
-            visitedTargets,
-          ) ??
-          ({
-            importedName: specifier.local.name,
-            resolvedFilePath: resolvedSourceFile,
-            sourceSpecifier: source.value,
-            isTypeOnly: specifierIsTypeOnly,
-            fromExportAll: false,
-          } satisfies ReExportTarget);
-
-        if (resolvedTarget.resolvedFilePath === barrelFilePath) {
-          return;
-        }
-
-        explicitReExports.set(getReExportKey(specifier.exported.name, specifierIsTypeOnly), {
-          importedName: resolvedTarget.importedName,
-          resolvedFilePath: resolvedTarget.resolvedFilePath,
-          sourceSpecifier: resolvedTarget.sourceSpecifier,
-          isTypeOnly: specifierIsTypeOnly,
-          fromExportAll: false,
-        });
-      });
-    }
-
-    if (
-      statement.type === AST_NODE_TYPES.ExportAllDeclaration &&
-      !statement.exported &&
-      statement.source !== null &&
-      statement.source.type === AST_NODE_TYPES.Literal &&
-      typeof statement.source.value === 'string'
-    ) {
-      const resolvedSourceFile = resolveImport(barrelFilePath, statement.source.value);
-
-      if (!resolvedSourceFile) {
-        return;
-      }
-
-      const exportedBindings = collectAllExportedBindings(resolvedSourceFile, resolveImport, new Set([barrelFilePath]));
-
-      exportedBindings.forEach(reExportKey => {
-        const { exportedName, isTypeOnly } = getReExportMeta(reExportKey);
-
-        if (statement.exportKind === 'type' && !isTypeOnly) {
-          return;
-        }
-
-        if (explicitReExports.has(reExportKey) || exportAllReExports.has(reExportKey)) {
-          return;
-        }
-
-        const resolvedTarget =
-          resolveReExportTarget(resolvedSourceFile, exportedName, isTypeOnly, resolveImport, visitedTargets) ??
-          ({
-            importedName: exportedName,
-            resolvedFilePath: resolvedSourceFile,
-            sourceSpecifier: statement.source.value,
-            isTypeOnly,
-            fromExportAll: true,
-          } satisfies ReExportTarget);
-
-        if (resolvedTarget.resolvedFilePath === barrelFilePath) {
-          return;
-        }
-
-        exportAllReExports.set(reExportKey, {
-          importedName: resolvedTarget.importedName,
-          resolvedFilePath: resolvedTarget.resolvedFilePath,
-          sourceSpecifier: resolvedTarget.sourceSpecifier,
-          isTypeOnly,
-          fromExportAll: true,
-        });
-      });
-    }
-  });
-
-  if (explicitReExports.size === 0 && exportAllReExports.size === 0) {
-    return null;
-  }
-
-  return {
-    explicitReExports,
-    exportAllReExports,
-  };
-}
-
-function parseBarrelFile(
-  barrelFilePath: string,
-  resolveImport: (importerFilename: string, specifier: string) => string | null,
-): BarrelAnalysis | null {
-  const program = parseModule(barrelFilePath);
-
-  if (!program) {
-    return null;
-  }
-
-  return collectBarrelAnalysis(program, barrelFilePath, resolveImport);
-}
 
 function isNamedImportSpecifier(specifier: TSESTree.ImportClause): specifier is NamedImportSpecifier {
   return (
@@ -496,189 +52,6 @@ function isNamedImportSpecifier(specifier: TSESTree.ImportClause): specifier is 
     specifier.imported.type === AST_NODE_TYPES.Identifier &&
     specifier.local.type === AST_NODE_TYPES.Identifier
   );
-}
-
-function isTypeOnlyImport(
-  declaration: TSESTree.ImportDeclaration,
-  specifier: TSESTree.ImportSpecifier | null = null,
-): boolean {
-  return declaration.importKind === 'type' || specifier?.importKind === 'type';
-}
-
-function serializeImportBinding(importedName: string, localName: string, isTypeOnly: boolean): string {
-  if (importedName === 'default') {
-    return `${isTypeOnly ? 'type ' : ''}default as ${localName}`;
-  }
-
-  const binding = importedName === localName ? importedName : `${importedName} as ${localName}`;
-
-  return isTypeOnly ? `type ${binding}` : binding;
-}
-
-function serializeImportBindings(
-  bindings: Array<{ importedName: string; isTypeOnly: boolean; localName: string }>,
-  allTypeOnly: boolean,
-): string {
-  return bindings
-    .map(binding =>
-      serializeImportBinding(binding.importedName, binding.localName, allTypeOnly ? false : binding.isTypeOnly),
-    )
-    .join(', ');
-}
-
-function getMergeableImportDeclaration(
-  node: TSESTree.ImportDeclaration,
-  sourceSpecifier: string,
-): TSESTree.ImportDeclaration | null {
-  if (node.source.value !== sourceSpecifier) {
-    return null;
-  }
-
-  if (
-    node.specifiers.some(
-      specifier =>
-        specifier.type !== AST_NODE_TYPES.ImportSpecifier ||
-        specifier.imported.type !== AST_NODE_TYPES.Identifier ||
-        specifier.local.type !== AST_NODE_TYPES.Identifier,
-    )
-  ) {
-    return null;
-  }
-
-  return node;
-}
-
-function buildAutofix(
-  sourceCode: TSESLint.SourceCode,
-  currentNode: TSESTree.ImportDeclaration,
-  matchedSpecifiers: MatchedSpecifier[],
-): TSESLint.ReportFixFunction | null {
-  const groupedImports = new Map<string, Array<{ importedName: string; isTypeOnly: boolean; localName: string }>>();
-
-  matchedSpecifiers.forEach(({ preferredSourceSpecifier, specifier, reExportTarget }) => {
-    if (!preferredSourceSpecifier) {
-      return;
-    }
-
-    const importBindings = groupedImports.get(preferredSourceSpecifier) ?? [];
-    importBindings.push({
-      importedName: reExportTarget.importedName,
-      isTypeOnly: reExportTarget.isTypeOnly,
-      localName: specifier.local.name,
-    });
-    groupedImports.set(preferredSourceSpecifier, importBindings);
-  });
-
-  const mergeTargets = new Map<string, TSESTree.ImportDeclaration>();
-
-  for (const sourceSpecifier of groupedImports.keys()) {
-    const matchingDeclarations = sourceCode.ast.body.filter(
-      statement =>
-        statement.type === AST_NODE_TYPES.ImportDeclaration &&
-        statement !== currentNode &&
-        getMergeableImportDeclaration(statement, sourceSpecifier),
-    ) as TSESTree.ImportDeclaration[];
-
-    if (matchingDeclarations.length > 1) {
-      return null;
-    }
-
-    if (matchingDeclarations.length === 1) {
-      mergeTargets.set(sourceSpecifier, matchingDeclarations[0]!);
-    }
-  }
-
-  return fixer => {
-    const fixes: TSESLint.RuleFix[] = [];
-
-    for (const [sourceSpecifier, bindings] of groupedImports.entries()) {
-      const mergeTarget = mergeTargets.get(sourceSpecifier);
-
-      if (!mergeTarget) {
-        continue;
-      }
-
-      const existingBindings = new Map<string, string>();
-      const existingTypeOnlyByLocalName = new Map<string, boolean>();
-      let hasConflict = false;
-
-      mergeTarget.specifiers.forEach(specifier => {
-        const importSpecifier = specifier as NamedImportSpecifier;
-        const localName = importSpecifier.local.name;
-        const importedName = importSpecifier.imported.name;
-        const typeOnly = isTypeOnlyImport(mergeTarget, importSpecifier);
-        const currentImportedName = existingBindings.get(localName);
-        const currentTypeOnly = existingTypeOnlyByLocalName.get(localName);
-
-        if (currentImportedName && (currentImportedName !== importedName || currentTypeOnly !== typeOnly)) {
-          hasConflict = true;
-          return;
-        }
-
-        existingBindings.set(localName, importedName);
-        existingTypeOnlyByLocalName.set(localName, typeOnly);
-      });
-
-      if (hasConflict) {
-        return null;
-      }
-
-      bindings.forEach(binding => {
-        const currentImportedName = existingBindings.get(binding.localName);
-        const currentTypeOnly = existingTypeOnlyByLocalName.get(binding.localName);
-
-        if (
-          currentImportedName &&
-          (currentImportedName !== binding.importedName || currentTypeOnly !== binding.isTypeOnly)
-        ) {
-          hasConflict = true;
-          return;
-        }
-
-        existingBindings.set(binding.localName, binding.importedName);
-        existingTypeOnlyByLocalName.set(binding.localName, binding.isTypeOnly);
-      });
-
-      if (hasConflict) {
-        return null;
-      }
-
-      const quote = mergeTarget.source.raw?.startsWith("'") ? "'" : '"';
-      const allTypeOnly = Array.from(existingTypeOnlyByLocalName.values()).every(Boolean);
-      const serializedBindings = serializeImportBindings(
-        Array.from(existingBindings.entries()).map(([localName, importedName]) => ({
-          importedName,
-          isTypeOnly: existingTypeOnlyByLocalName.get(localName)!,
-          localName,
-        })),
-        allTypeOnly,
-      );
-      const mergedImport = `import${allTypeOnly ? ' type' : ''} { ${serializedBindings} } from ${quote}${sourceSpecifier}${quote};`;
-
-      fixes.push(fixer.replaceText(mergeTarget, mergedImport));
-      groupedImports.delete(sourceSpecifier);
-    }
-
-    const quote = currentNode.source.raw?.startsWith("'") ? "'" : '"';
-    const replacement = Array.from(groupedImports.entries())
-      .map(([sourceSpecifier, bindings]) => {
-        const allTypeOnly = bindings.every(binding => binding.isTypeOnly);
-        const serializedBindings = serializeImportBindings(bindings, allTypeOnly);
-
-        return `import${allTypeOnly ? ' type' : ''} { ${serializedBindings} } from ${quote}${sourceSpecifier}${quote};`;
-      })
-      .join('\n');
-
-    if (replacement) {
-      fixes.push(fixer.replaceText(currentNode, replacement));
-    } else {
-      const [start, end] = currentNode.range;
-      const adjustedEnd = sourceCode.text.slice(end, end + 1) === '\n' ? end + 1 : end;
-      fixes.push(fixer.removeRange([start, adjustedEnd]));
-    }
-
-    return fixes;
-  };
 }
 
 function getTsconfigPath(
